requirements.txt
aiogram==3.3.0
python-dotenv
rembg
ffmpeg-python
onnxruntime
